requirements.txt
Numpy
matplotlib
tifffile
pathlib
pandas
random2
pillow
scipy

